add session to admin function
soo user cant access to admin function

+++ hapus.php
@@ -1,4 +1,9 @@
-<?php 
+<?php
+session_start();
+if (!isset($_SESSION['user']) || !isset($_SESSION['role']) || $_SESSION['role'] !== 'admin') {
+    header("Location: ../index.php");
+    exit;
+}
 include "koneksi/db.php"; 
 $id = $_GET['id']; 
 mysqli_query($conn, "DELETE FROM users WHERE id=$id");
